+++ lib/src/core/service.dart
@@ -4,12 +4,12 @@ import 'dart:async';
 import 'package:angel_http_exception/angel_http_exception.dart';
 import 'package:merge_map/merge_map.dart';
 import '../util.dart';
-import 'angel_base.dart';
 import 'hooked_service.dart' show HookedService;
 import 'metadata.dart';
 import 'request_context.dart';
 import 'response_context.dart';
 import 'routable.dart';
+import 'server.dart';
 
 /// Indicates how the service was accessed.
 ///
@@ -59,7 +59,7 @@ class Service extends Routable {
   List get bootstrappers => [];
 
   /// The [Angel] app powering this service.
-  AngelBase app;
+  Angel app;
 
   /// Closes this service, including any database connections or stream controllers.
   void close() {}