+++ CHANGELOG.md
@@ -10,3 +10,4 @@
 * Remove `contentType` argument in `ResponseContext.serialize`.
 * Moved `lib/hooks.dart` into `package:angel_hooks`.
 * Moved `TypedService` into `package:angel_typed_service`.
+* Completely removed the `AngelBase` class.